Crate akka_projection_rs

Akka projections

In Akka Projections you process a stream of event envelopes from a source to a projected model or external system. Each envelope is associated with an offset representing the position in the stream. This offset is used for resuming the stream from that position when the projection is restarted.

Modules

consumer

Akka projections

consumer_filter

The consumer may define declarative filters that are sent to the producer and evaluated on the producer side before emitting the events.

offset_store

An offset store keeps track of the last offset used in relation to an entity, including its entity type. It can be used in various places, for example, when consuming event envelopes from a remote projection, where they be emitted by the consumer out of order. This entity can track such situations.

Structs

FnHandler

HandlerError

Errors for event processing by a handler.

SourceProviderError

Errors for event processing by a handler.

UnusedHandler

For the purposes of constructing unused handlers.

UnusedPendingHandler

For the purposes of constructing unused handlers.

Enums

Handlers

Captures the various types of handlers and the way they are performed.

Traits

Handler

Handle event envelopes in any way that an application requires.

PendingHandler

Handle event envelopes in any way that an application requires.

SourceProvider

Provides a source of envelopes using slices as a query.