Enum akka_projection_rs::consumer_filter::FilterCriteria

pub enum FilterCriteria {

    ExcludeTags {
        tags: Vec<Tag>,
    },
    RemoveExcludeTags {
        tags: Vec<Tag>,
    },
    IncludeTags {
        tags: Vec<Tag>,
    },
    RemoveIncludeTags {
        tags: Vec<Tag>,
    },
    ExcludeRegexEntityIds {
        matching: Vec<ComparableRegex>,
    },
    RemoveExcludeRegexEntityIds {
        matching: Vec<ComparableRegex>,
    },
    IncludeRegexEntityIds {
        matching: Vec<ComparableRegex>,
    },
    RemoveIncludeRegexEntityIds {
        matching: Vec<ComparableRegex>,
    },
    ExcludeEntityIds {
        entity_ids: Vec<EntityId>,
    },
    RemoveExcludeEntityIds {
        entity_ids: Vec<EntityId>,
    },
    IncludeEntityIds {
        entity_id_offsets: Vec<EntityIdOffset>,
    },
    RemoveIncludeEntityIds {
        entity_ids: Vec<EntityId>,
    },
    IncludeTopics {
        expressions: Vec<TopicMatcher>,
    },
    RemoveIncludeTopics {
        expressions: Vec<TopicMatcher>,
    },

}

Exclude criteria are evaluated first. If no matching exclude criteria the event is emitted. If an exclude criteria is matching the include criteria are evaluated. If no matching include criteria the event is discarded. If matching include criteria the event is emitted.

Variants

ExcludeTags

Fields

tags: Vec<Tag>

Exclude events with any of the given tags, unless there is a matching include filter that overrides the exclude.

RemoveExcludeTags

Fields

tags: Vec<Tag>

Remove a previously added ExcludeTags.

IncludeTags

Fields

tags: Vec<Tag>

Include events with any of the given tags. A matching include overrides a matching exclude.

RemoveIncludeTags

Fields

tags: Vec<Tag>

Remove a previously added IncludeTags.

ExcludeRegexEntityIds

Fields

matching: Vec<ComparableRegex>

Exclude events for entities with entity ids matching the given regular expressions, unless there is a matching include filter that overrides the exclude.

RemoveExcludeRegexEntityIds

Fields

matching: Vec<ComparableRegex>

Remove a previously added ExcludeRegexEntityIds.

IncludeRegexEntityIds

Fields

matching: Vec<ComparableRegex>

Include events for entities with entity ids matching the given regular expressions. A matching include overrides a matching exclude.

RemoveIncludeRegexEntityIds

Fields

matching: Vec<ComparableRegex>

Remove a previously added IncludeRegexEntityIds.

ExcludeEntityIds

Fields

entity_ids: Vec<EntityId>

Exclude events for entities with the given entity ids, unless there is a matching include filter that overrides the exclude.

RemoveExcludeEntityIds

Fields

entity_ids: Vec<EntityId>

Remove a previously added ExcludeEntityIds.

IncludeEntityIds

Fields

entity_id_offsets: Vec<EntityIdOffset>

Include events for entities with the given entity ids. A matching include overrides a matching exclude.

For the given entity ids a seq_nr can be defined to replay all events for the entity from the sequence number (inclusive). If seq_nr is 0 events will not be replayed.

RemoveIncludeEntityIds

Fields

entity_ids: Vec<EntityId>

Remove a previously added IncludeEntityIds.

IncludeTopics

Fields

expressions: Vec<TopicMatcher>

Include events with any of the given matching topics. A matching include overrides a matching exclude.

RemoveIncludeTopics

Fields

expressions: Vec<TopicMatcher>

Remove a previously added IncludeTopics.